requirements.txt
azure-functions
requests       
tiktoken
azure-cognitiveservices-speech
python-dotenv
pyodbc
azure-identity
azure-storage-blob